+++ tsdb/engine/tsm1/iterator.go
@@ -54,3 +54,38 @@ func (c cursorsAt) close() {
 		cur.close()
 	}
 }
+
+// newMergeFinalizerIterator creates a new Merge iterator from the inputs. If the call to Merge succeeds,
+// the resulting Iterator will be wrapped in a finalizer iterator.
+// If Merge returns an error, the inputs will be closed.
+func newMergeFinalizerIterator(inputs []influxql.Iterator, opt influxql.IteratorOptions) (influxql.Iterator, error) {
+	itr, err := influxql.Iterators(inputs).Merge(opt)
+	if err != nil {
+		influxql.Iterators(inputs).Close()
+		return nil, err
+	}
+	return newFinalizerIterator(itr), nil
+}
+
+// newFinalizerIterator creates a new iterator that installs a runtime finalizer
+// to ensure close is eventually called if the iterator is garbage collected.
+// This additional guard attempts to protect against clients of CreateIterator not
+// correctly closing them and leaking cursors.
+func newFinalizerIterator(itr influxql.Iterator) influxql.Iterator {
+	if itr == nil {
+		return nil
+	}
+
+	switch inner := itr.(type) {
+	case influxql.FloatIterator:
+		return newFloatFinalizerIterator(inner)
+	case influxql.IntegerIterator:
+		return newIntegerFinalizerIterator(inner)
+	case influxql.StringIterator:
+		return newStringFinalizerIterator(inner)
+	case influxql.BooleanIterator:
+		return newBooleanFinalizerIterator(inner)
+	default:
+		panic(fmt.Sprintf("unsupported finalizer iterator type: %T", itr))
+	}
+}